refactor(clay-charts): remove indiviual exported chart types
